+++ test/blackbox-tests/test-cases/describe-pp/describe-staged-pps.t/src/main.ml
@@ -1,1 +1,0 @@
-let () = print_string [%add_suffix "hello"]